// src/app/utils/enums.ts
export enum TIPO_MESSAGGI {
    SUCCESS = 'SUCCESS',
    WARNING = 'WARNING',
    ERROR = 'ERRORE'
}

export enum RUOLI {
    AMMINISTRATORE = 6,
    CORDINATOR = 7,
    NG_EXPERT = 8,
    EESSI_MEMBER = 9,
    SUPERVISOR_CE = 10,
    NG_PATNER = 11,
    EESSI_EVENT_MANAGER = 16,
    HELP_DESK = 17,
    CONTENT_MANAGER = 18,
    EESSI_MEMBER_PLUS = 19,
    RINA_HANDOVER_MEMBER = 20,
    RINA_HANDOVER_MEMBER_PLUS = 21,
}

export enum STATO_DOCUMENTI {
    BOZZA = 0,
    DA_APPROVARE = 1,
    UFFICIALE = 2
}